requirements.txt
discord.py==2.1.1
requests==2.28.2
Commitizen==2.42.1